gh-76785: Fixes for test.support.interpreters (gh-112982)
This involves a number of changes for PEP 734.
