+++ Lib/test/test_interpreters/__main__.py
@@ -1,0 +1,4 @@
+from . import load_tests
+import unittest
+
+nittest.main()
